+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v2.2.0 - 2022-10-30
+
+### Added
+
+- `new()` and `new_string()`
+
+### Deprecated
+
+- `scru128()` and `scru128_string()` to promote `scru128.new()` syntax over
+  `from scru128 import scru128`
+
+### Maintenance
+
+- Updated dev dependencies
+
 ## v2.1.2 - 2022-06-11
 
 ### Fixed